+++ messages/nl/admin.php
@@ -1,14 +1,9 @@
 <?php
-return array (
-  '<strong>Dark Mode</strong> module configuration' => 'Moduleconfiguratie <strong>Donkere modus</strong>',
-  'Adds a Dark Mode to HumHub' => 'Voegt een donkere modus toe aan HumHub',
-  'Current base theme: ' => 'Huidig basisthema:',
-  'Dark Mode' => 'Donker modus',
-  'Dark Theme' => 'Donkerthema',
-  'Donate' => 'Doneer',
-  'GitHub Repository' => 'GitHub Repository',
-  'Guide for Administrators' => 'Handleiding voor beheerders',
-  'Show Button in Top Bar' => 'Knop weergeven in bovenste balk',
-  'The stylesheet of the selected theme will be used for the dark mode.' => 'Voor de donkere modus wordt het stylesheet van het geselecteerde thema gebruikt.',
-  'Users can set their theme preferences also in Account Settings > General.' => 'Gebruikers kunnen hun themavoorkeuren ook instellen in Accountinstellingen &gt; Algemeen.',
-);
+
+return [
+    '<strong>Dark Mode</strong> module configuration' => 'Moduleconfiguratie <strong>Donkere modus</strong>',
+    'Dark Mode' => 'Donker modus',
+    'Show Button in Top Bar' => 'Knop weergeven in bovenste balk',
+    'Users can set their theme preferences also in Account Settings > General.' => 'Gebruikers kunnen hun themavoorkeuren ook instellen in Accountinstellingen &gt; Algemeen.',
+    'Adds a Dark Mode Switcher to HumHub' => '',
+];