+++ messages/sv/admin.php
@@ -1,14 +1,9 @@
 <?php
+
 return [
-  '<strong>Dark Mode</strong> module configuration' => '<strong>Mörkt läge</strong>-modulkonfiguration',
-  'Adds a Dark Mode to HumHub' => 'Lägger till ett mörkt läge till HumHub',
-  'Current base theme: ' => 'Nuvarande grundtema:',
-  'Dark Mode' => 'Mörkt läge',
-  'Dark Theme' => 'Mörkt tema',
-  'Donate' => 'Donera',
-  'GitHub Repository' => 'GitHub Repository',
-  'Guide for Administrators' => 'Guide för administratörer',
-  'Show Button in Top Bar' => 'Visa knapp i toppfältet',
-  'The stylesheet of the selected theme will be used for the dark mode.' => 'Stilmallen för det valda temat kommer att användas för mörkt läge.',
-  'Users can set their theme preferences also in Account Settings > General.' => 'Användare kan även ställa in sina temapreferenser i Kontoinställningar &gt; Allmänt.',
+    '<strong>Dark Mode</strong> module configuration' => '<strong>Mörkt läge</strong>-modulkonfiguration',
+    'Dark Mode' => 'Mörkt läge',
+    'Show Button in Top Bar' => 'Visa knapp i toppfältet',
+    'Users can set their theme preferences also in Account Settings > General.' => 'Användare kan även ställa in sina temapreferenser i Kontoinställningar &gt; Allmänt.',
+    'Adds a Dark Mode Switcher to HumHub' => '',
 ];