+++ models/Config.php
@@ -2,8 +2,6 @@
 
 namespace humhub\modules\darkMode\models;
 
-use humhub\modules\darkMode\assets\DarkStyleAsset;
-use humhub\modules\ui\view\helpers\ThemeHelper;
 use Yii;
 
 /**
@@ -14,8 +12,6 @@ class Config extends \yii\base\Model
     const DARK_CSS_SUFFIX = ' (dark)';
     const FALLBACK = 'DarkHumHub';
 
-    public $theme;
-
     public $showButton = true;
 
     public function init()
@@ -25,23 +21,14 @@ class Config extends \yii\base\Model
         $module = Yii::$app->getModule('dark-mode');
         // make sure module is enabled before retrieving settings, see https://github.com/felixhahnweilheim/humhub-dark-mode/issues/48
         if ($module) {
-            $this->theme = $module->settings->get('theme');
             $this->showButton = $module->settings->get('showButton', $this->showButton);
         }
 
-        // If no setting was found, get recommended theme or fallback (DarkHumHub) 
-        if (empty($this->theme)) {
-            $this->theme = self::getRecommendedTheme();
-            if (empty($this->theme)) {
-                $this->theme = self::FALLBACK;
-            }
-        }
     }
 
     public function rules()
     {
         return [
-            ['theme', 'in', 'range' => array_keys($this->getThemes())],
             ['showButton', 'boolean']
         ];
     }
@@ -49,7 +36,6 @@ class Config extends \yii\base\Model
     public function attributeLabels()
     {
         return [
-            'theme' => Yii::t('DarkModeModule.admin', 'Dark Theme'),
             'showButton' => Yii::t('DarkModeModule.admin', 'Show Button in Top Bar')
         ];
     }
@@ -57,35 +43,8 @@ class Config extends \yii\base\Model
     public function attributeHints()
     {
         return [
-            'theme' => Yii::t('DarkModeModule.admin', 'The stylesheet of the selected theme will be used for the dark mode.'),
             'showButton' => Yii::t('DarkModeModule.admin', 'Users can set their theme preferences also in Account Settings > General.')
         ];
-    }
-
-    public function getThemes()
-    {
-        $themes = [];
-
-        foreach (ThemeHelper::getThemes() as $theme) {
-            // Themes with a dark.css
-            if (file_exists($theme->basePath . DIRECTORY_SEPARATOR . 'css' . DIRECTORY_SEPARATOR . 'dark.css')) {
-                $themes[$theme->name . self::DARK_CSS_SUFFIX] = $theme->name . self::DARK_CSS_SUFFIX;
-                // Themes containing "dark" in their name
-            } elseif (stripos($theme->name, 'dark') !== false) {
-                $themes[$theme->name] = $theme->name;
-            }
-        }
-
-        // Add "HumHub (dark)"
-        $themes['DarkHumHub'] = 'HumHub' . self::DARK_CSS_SUFFIX;
-
-        // Add "enterprise (dark)" if module enabled
-        $enterprise = 'enterprise-theme';
-        if (Yii::$app->hasModule($enterprise) && isset(Yii::$app->modules[$enterprise])) {
-            $themes['DarkEnterprise'] = 'enterprise' . self::DARK_CSS_SUFFIX;
-        }
-
-        return $themes;
     }
 
     public function save()
@@ -95,55 +54,8 @@ class Config extends \yii\base\Model
         }
 
         $settings = Yii::$app->getModule('dark-mode')->settings;
-        $settings->set('theme', $this->theme);
         $settings->set('showButton', $this->showButton);
 
-        Yii::$app->cache->delete(DarkStyleAsset::PATH_CACHE);
-        Yii::$app->cache->delete(DarkStyleAsset::FILENAME_CACHE);
-
         return true;
-    }
-
-    /*
-     * returns array path and file name of stylesheet
-     * used for the assets
-     */
-    public function getThemeInfos()
-    {
-        $info['fileName'] = 'theme.css';
-
-        if ($this->theme == 'DarkEnterprise') {
-            $info['path'] = '@dark-mode/resources/DarkEnterprise/css';
-        } elseif ($this->theme == 'DarkHumHub') {
-            $info['path'] = '@dark-mode/resources/DarkHumHub/css';
-        } else {
-            // Themes with dark.css
-            if (strpos($this->theme, self::DARK_CSS_SUFFIX) !== false) {
-                $this->theme = str_replace(self::DARK_CSS_SUFFIX, '', $this->theme);
-                $info['fileName'] = 'dark.css';
-            }
-            $info['path'] = ThemeHelper::getThemeByName($this->theme)->basePath . DIRECTORY_SEPARATOR . 'css';
-        }
-        return $info;
-    }
-
-    // Try to return recommended theme
-    public function getRecommendedTheme()
-    {
-        $baseTheme = ThemeHelper::getThemeByName(Yii::$app->view->theme->name);
-
-        if (!$baseTheme || $baseTheme->name === 'HumHub') {
-            return 'DarkHumHub';
-        }
-
-        if ($baseTheme->name === 'enterprise') {
-            return 'DarkEnterprise';
-        }
-
-        if (file_exists($baseTheme->basePath . DIRECTORY_SEPARATOR . 'css' . DIRECTORY_SEPARATOR . 'dark.css')) {
-            return $baseTheme->name . self::DARK_CSS_SUFFIX;
-        }
-
-        return '';
     }
 }